+++ src/messages/es/admin.php
@@ -522,4 +522,5 @@ return [
     'i18n_required_validator_invalid_format' => 'El atributo dado \"{attribute}\" tiene que ser de tipo array.',
     'i18n_required_validator_missing_key' => 'La key del idioma \"{lang}\" no se encuentra y es requerida.',
     'i18n_required_validator_invalid_empty_value' => 'El valor del idioma \"{lang}\" no puede estar vacío.',
+    'crud_exportdata_col_filter' => 'Filter',
 ];